Fix 32bit build error of System.Net.Security.Native
By using size_t rather than uint64_t for sizes,
we can ensure that it is 32b in 32b systems and
it's 64b in 64b systems.

Fix #7090

+++ src/Native/System.Net.Security.Native/pal_gssapi.cpp
@@ -232,7 +232,7 @@ extern "C" void NetSecurityNative_ReleaseGssBuffer(void* buffer, uint64_t length
     assert(buffer != nullptr);
 
     uint32_t minorStatus;
-    GssBuffer gssBuffer{.length = length, .value = buffer};
+    GssBuffer gssBuffer{.length = static_cast<size_t>(length), .value = buffer};
     gss_release_buffer(&minorStatus, &gssBuffer);
 }
